requirements.txt
alembic==1.7.7
apturl==0.5.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
astroid==2.9.3
asttokens==2.0.5
attrs==21.4.0
autopep8==1.6.0
awscli==1.22.34
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.11.1
bleach==5.0.0
blinker==1.4
botocore==1.23.34
Brlapi==0.8.3
certifi==2020.6.20
cffi==1.15.0
chardet==4.0.0
cliapp==1.20180812.1
click==7.1.2
cmdtest==0.32+git
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
debugpy==1.6.0
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
distro==1.7.0
distro-info===1.1build1
Django==4.0.6
dnspython==2.2.1
docutils==0.17.1
dominate==2.6.0
duplicity==0.8.21
ecdsa==0.17.0
email-validator==1.2.1
entrypoints==0.4
executing==0.8.3
Faker==13.11.0
fasteners==0.14.1
fastjsonschema==2.15.3
flasgger==0.9.5
Flask==1.1.4
Flask-Autodoc==0.1.2
Flask-Bootstrap==3.3.7.1
Flask-Cors==3.0.10
Flask-Mail==0.9.1
Flask-Migrate==2.6.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
Flask-UUID==0.2
future==0.18.2
greenlet==1.1.2
gyp==0.1
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
ipykernel==6.13.1
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
isort==5.6.4
itsdangerous==1.1.0
jedi==0.18.1
jeepney==0.7.1
Jinja2==2.11.3
jmespath==0.10.0
jsonschema==4.5.1
jupyter==1.0.0
jupyter-client==7.3.3
jupyter-console==6.4.3
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
kazam==1.4.5
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy-object-proxy==0.0.0
lockfile==0.12.2
logilab-common==1.8.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mccabe==0.6.1
mistune==2.0.2
monotonic==1.6
more-itertools==8.10.0
mypy-extensions==0.4.3
nbclient==0.6.4
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
netifaces==0.11.0
notebook==6.4.11
oauthlib==3.2.0
olefile==0.46
packaging==21.3
pandocfilters==1.5.0
paramiko==2.9.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.5.1
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==3.12.4
psutil==5.9.1
psycopg2-binary==2.8.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pycairo==1.20.1
pycodestyle==2.8.0
pycparser==2.21
pycups==2.0.1
Pygments==2.11.2
PyGObject==3.42.0
PyJWT==2.3.0
pylint==2.12.2
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.18.1
python-apt==2.3.0+ubuntu2
python-dateutil==2.8.1
python-debian===0.1.43ubuntu1
python-dotenv==0.20.0
python-jose==3.3.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==23.1.0
qtconsole==5.3.1
QtPy==2.1.0
reportlab==3.6.8
requests==2.25.1
roman==3.3
rsa==4.8
s3transfer==0.5.0
SecretStorage==3.3.1
Send2Trash==1.8.0
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.36
sqlparse==0.4.2
stack-data==0.2.0
systemd-python==234
terminado==0.15.0
tinycss2==1.1.1
toml==0.10.2
tornado==6.1
traitlets==5.2.2.post1
ttystatus==0.38
typing-extensions==3.10.0.2
ubuntu-advantage-tools==27.9
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
visitor==0.1.3
wadllib==1.3.6
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.6.0
wrapt==1.13.3
xdg==5
xkit==0.0.0
zipp==1.0.0
